Split core package into multiple files
