+++ core/doc.go
@@ -1,0 +1,2 @@
+// Package core implements basic chess functionality.
+package core
